Re-generate components to fix version type

+++ mcu-firmware/rrrc/components/MotorPortHandler/MotorPortHandler.h
@@ -37,7 +37,7 @@ typedef union {
 } DriveRequest_RequestValue_t;
 
 typedef struct {
-    uint32_t version;
+    uint8_t version;
     float power_limit;
     float speed_limit;
     DriveRequest_RequestType_t request_type;
